// src/app/core/api/generated/schemas/inventoryDetailDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { InventoryDetailDtoInventorySource } from "./inventoryDetailDtoInventorySource";
import type { InventoryDetailDtoLeadTimeDays } from "./inventoryDetailDtoLeadTimeDays";
import type { InventoryDetailDtoLowStockThreshold } from "./inventoryDetailDtoLowStockThreshold";
import type { InventoryDetailDtoSku } from "./inventoryDetailDtoSku";
import type { InventoryDetailDtoSlug } from "./inventoryDetailDtoSlug";
import type { InventoryDetailDtoTypeName } from "./inventoryDetailDtoTypeName";
import type { InventoryDetailDtoWarehouseBin } from "./inventoryDetailDtoWarehouseBin";

export interface InventoryDetailDto {
  product_id: number;
  title: string;
  slug?: InventoryDetailDtoSlug;
  /** @nullable */
  type_name?: InventoryDetailDtoTypeName;
  /** @nullable */
  sku?: InventoryDetailDtoSku;
  selling_price: number;
  available_qty: number;
  sold_qty: number;
  updated_at: string;
  status: string;
  stock_status: string;
  low_stock: boolean;
  /** @nullable */
  low_stock_threshold?: InventoryDetailDtoLowStockThreshold;
  /** @nullable */
  inventory_source?: InventoryDetailDtoInventorySource;
  /** @nullable */
  warehouse_bin?: InventoryDetailDtoWarehouseBin;
  /** @nullable */
  lead_time_days?: InventoryDetailDtoLeadTimeDays;
  is_active: boolean;
}