// src/app/core/api/generated/schemas/inventoryListItemDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { InventoryListItemDtoLowStockThreshold } from "./inventoryListItemDtoLowStockThreshold";
import type { InventoryListItemDtoSku } from "./inventoryListItemDtoSku";
import type { InventoryListItemDtoSlug } from "./inventoryListItemDtoSlug";
import type { InventoryListItemDtoTypeName } from "./inventoryListItemDtoTypeName";

export interface InventoryListItemDto {
  product_id: number;
  title: string;
  slug?: InventoryListItemDtoSlug;
  /** @nullable */
  type_name?: InventoryListItemDtoTypeName;
  /** @nullable */
  sku?: InventoryListItemDtoSku;
  selling_price: number;
  available_qty: number;
  sold_qty: number;
  updated_at: string;
  status: string;
  stock_status: string;
  low_stock: boolean;
  /** @nullable */
  low_stock_threshold?: InventoryListItemDtoLowStockThreshold;
}