// src/app/core/api/generated/schemas/updateInventoryDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */

export interface UpdateInventoryDto {
  stock_qty?: number;
  low_stock_threshold?: number;
  stock_status?: string;
  status?: string;
  is_active?: boolean;
  inventory_source?: string;
  warehouse_bin?: string;
  lead_time_days?: number;
}